setting default log directory

+++ main.py
@@ -1,35 +1,66 @@
-from tkinter import *
+import tkinter as tk
 from tkinter import filedialog
+from tkinter import messagebox
 import os
+import os.path
 from tkinter import ttk
 import json
 
-# getting directory
+
+# set directory of logs trough explorer
+def set_dir():
+    tk.messagebox.showinfo(title="Log Directory",
+                           message="select your chatterino log folder, e.g.: "
+                                   "\"C:/.../Chatterino2/Logs/Twitch\"")
+
+    root.filename = filedialog.askdirectory()
+
+    if root.filename != "":
+        with open('directory.json', 'w') as f:
+            json.dump(str(root.filename), f)
+
+        return str(root.filename)
+    else:
+        return get_dir()
+
+
+# get the saved directory from json file
+def get_dir():
+    with open('directory.json') as f:
+        log_dir_json = json.load(f)
+    print(log_dir_json)
+    return log_dir_json
+
+
+# get all log folders within log dir
+def get_folders():
+    folders = []
+    os.walk(get_dir())
+    print(folders)
+
+##################################################################################################################
+
+
+# checking if dir json file exists, else creating default file
+if not os.path.isfile("directory.json"):
+    with open('directory.json', 'w') as f:
+        json.dump(str(os.getenv('APPDATA') + "\\Chattterino2\\Logs\\Twitch"), f)
+
+
+# setting log dir
+log_dir = get_dir()
 
 
 # set up window
-root = Tk()
+root = tk.Tk()
 root.title("search logs")
 root.configure(bg="#424242")
 
 
-def set_dir():
-    root.filename = filedialog.askdirectory()
-    print(root.filename)
-
-    with open('directory.json', 'w') as f:
-        json.dump(str(root.filename), f)
-
-    return str(root.filename)
+##################################################################################################################
 
 
-def get_dir():
-    with open('directory.json') as f:
-        data = json.load(f)
-    print(data)
-
-
-channel = StringVar()
+channel = tk.StringVar()
 channel_chosen = ttk.Combobox(root, width=25, textvariable=channel)
 
 # Adding combobox drop down list
@@ -47,11 +78,11 @@ channel_chosen['values'] = (' January',
                             ' December')
 
 # labels
-label_channel = Label(root, text="Channel: ", bg="#424242", fg="white")
+label_channel = tk.Label(root, text="Channel: ", bg="#424242", fg="white")
 
 # buttons
-button_dir = Button(root, text="set directory", command=lambda: set_dir())
-button_get_dir = Button(root, text="get directory", command=lambda: get_dir())
+button_dir = tk.Button(root, text="change directory", command=lambda: set_dir())
+button_get_dir = tk.Button(root, text="get directory", command=lambda: get_dir())
 
 # grid
 label_channel.grid(row=0, column=0)
